[Refactor] add `isParenthesized` AST util
