+++ lib/rules/jsx-wrap-multilines.js
@@ -9,6 +9,7 @@ const has = require('object.hasown/polyfill')();
 const docsUrl = require('../util/docsUrl');
 const jsxUtil = require('../util/jsx');
 const reportC = require('../util/report');
+const isParenthesized = require('../util/ast').isParenthesized;
 
 // ------------------------------------------------------------------------------
 // Constants
@@ -89,20 +90,10 @@ module.exports = {
       return option && option !== 'ignore';
     }
 
-    function isParenthesised(node) {
-      const sourceCode = context.getSourceCode();
-      const previousToken = sourceCode.getTokenBefore(node);
-      const nextToken = sourceCode.getTokenAfter(node);
-
-      return previousToken && nextToken
-        && previousToken.value === '(' && previousToken.range[1] <= node.range[0]
-        && nextToken.value === ')' && nextToken.range[0] >= node.range[1];
-    }
-
     function needsOpeningNewLine(node) {
       const previousToken = context.getSourceCode().getTokenBefore(node);
 
-      if (!isParenthesised(node)) {
+      if (!isParenthesized(context, node)) {
         return false;
       }
 
@@ -116,7 +107,7 @@ module.exports = {
     function needsClosingNewLine(node) {
       const nextToken = context.getSourceCode().getTokenAfter(node);
 
-      if (!isParenthesised(node)) {
+      if (!isParenthesized(context, node)) {
         return false;
       }
 
@@ -153,12 +144,12 @@ module.exports = {
       const sourceCode = context.getSourceCode();
       const option = getOption(type);
 
-      if ((option === true || option === 'parens') && !isParenthesised(node) && isMultilines(node)) {
+      if ((option === true || option === 'parens') && !isParenthesized(context, node) && isMultilines(node)) {
         report(node, 'missingParens', (fixer) => fixer.replaceText(node, `(${sourceCode.getText(node)})`));
       }
 
       if (option === 'parens-new-line' && isMultilines(node)) {
-        if (!isParenthesised(node)) {
+        if (!isParenthesized(context, node)) {
           const tokenBefore = sourceCode.getTokenBefore(node, { includeComments: true });
           const tokenAfter = sourceCode.getTokenAfter(node, { includeComments: true });
           const start = node.loc.start;
