+++ lib/util/ast.js
@@ -295,6 +295,23 @@ function getKeyValue(context, node) {
 }
 
 /**
+ * Checks if a node is surrounded by parenthesis.
+ *
+ * @param {object} context - Context from the rule
+ * @param {ASTNode} node - Node to be checked
+ * @returns {boolean}
+ */
+function isParenthesized(context, node) {
+  const sourceCode = context.getSourceCode();
+  const previousToken = sourceCode.getTokenBefore(node);
+  const nextToken = sourceCode.getTokenAfter(node);
+
+  return !!previousToken && !!nextToken
+    && previousToken.value === '(' && previousToken.range[1] <= node.range[0]
+    && nextToken.value === ')' && nextToken.range[0] >= node.range[1];
+}
+
+/**
  * Checks if a node is being assigned a value: props.bar = 'bar'
  * @param {ASTNode} node The AST node being checked.
  * @returns {Boolean}
@@ -410,6 +427,7 @@ module.exports = {
   getPropertyNameNode,
   getComponentProperties,
   getKeyValue,
+  isParenthesized,
   isAssignmentLHS,
   isClass,
   isFunction,
